Remove 'play' on visualisation of locations
