+++ index.html
@@ -117,6 +117,7 @@
     <script type="module" src="views/oneNameTrees/locationstatistics.js"></script>
     <script type="module" src="views/oneNameTrees/familytreestatistics.js"></script>
     <script type="module" src="views/oneNameTrees/oneNameTrees.js"></script>
+    <script type="module" src="views/oneNameTrees/D3DataFormatter.js"></script>
 
     <link rel="stylesheet" href="views/oneNameTrees/oneNameTrees.css" />
   </head>
